backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
browser-use==0.1.0
python-dotenv==1.0.0
pandas==2.1.3
pydantic==2.5.2
pytest==7.4.3
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiosmtplib==2.0.2
jinja2==3.1.2 